Update xkcdViewer provider to use HTTPS URLs
xkcd.com changed to using HTTPS some time ago; the change causes
XkcdViewer to fail with "Data returned from website didn't match
expected format" as the 301 redirect is attempted to be parsed as JSON,
which obviously fails.

While this isn't a complete fix, it should hopefully fix the app enough
for it to work again.

Fixes #70.